Исправлен Readme,вынесен Config в отдельный класс BaseOrjsonModel, обавлен workflows на проверку flake8 (нестрогий)
