+++ README.md
@@ -1,13 +1,24 @@
-# Проектная работа 4 спринта
-
+# Проект Async_API
 Ссылка на репозиторий - https://github.com/ManulKyn/Async_API_sprint_1
 
-**Важное сообщение для тимлида:** для ускорения проверки проекта укажите ссылку на приватный репозиторий с командной работой в файле readme и отправьте свежее приглашение на аккаунт [BlueDeep](https://github.com/BigDeepBlue).
+![Async_API workflow](https://github.com/ManulKyn/Async_API_sprint_1/actions/workflows/Async_API.yml/badge.svg)
 
-В папке **tasks** ваша команда найдёт задачи, которые необходимо выполнить в первом спринте второго модуля.  Обратите внимание на задачи **00_create_repo** и **01_create_basis**. Они расцениваются как блокирующие для командной работы, поэтому их необходимо выполнить как можно раньше.
+В данном проекте использовались такие технологии как GitHub, Docker, NGINX, Python, Elasticsearch, PostgreSQL, Redis
 
-Мы оценили задачи в стори поинтах, значения которых брались из [последовательности Фибоначчи](https://ru.wikipedia.org/wiki/Числа_Фибоначчи) (1,2,3,5,8,…).
+# Установка проекта
+1. Клонируете проект на локальную машину (git clone https://github.com/ManulKyn/Async_API_sprint_1.git) 
+2. Развертывание проекта из основной папки (docker-compose up --build -d)
 
-Вы можете разбить имеющиеся задачи на более маленькие, например, распределять между участниками команды не большие куски задания, а маленькие подзадачи. В таком случае не забудьте зафиксировать изменения в issues в репозитории.
+# Не много о проекте и его назначении
+Дипломный проект, в котором разворачивается ETL по переливке данных из БД Postgre в Elasticsearch и создании индексов и сервис с FastAPI, который кэширует запросы в Redis.
 
-**От каждого разработчика ожидается выполнение минимум 40% от общего числа стори поинтов в спринте.**
+# Развернутый сервер можно посмотреть по адресу:
+http://127.0.0.1:8000/ - главная страница
+http://127.0.0.1:8000/api/openapi#/ - доступные fastapi
+
+# Инфо об авторах проекта 
+Салынов Андрей, Студент факультета мидл-питон. Когорта 15. 
+Другие репозитории https://github.com/ManulKyn
+
+Байрачный Даниил, Студент факультета мидл-питон. Когорта  15.
+Другие репозитории https://github.com/bdannyv
