+++ service/models/base.py
@@ -1,0 +1,12 @@
+import orjson
+from pydantic import BaseModel
+
+
+def orjson_dumps(v, *, default):
+    return orjson.dumps(v, default=default).decode()
+
+
+class BaseOrjsonModel(BaseModel):
+    class Config:
+        json_loads = orjson.loads
+        json_dumps = orjson_dumps
