+++ service/models/genre.py
@@ -1,19 +1,12 @@
 from typing import Optional
 
-import orjson
 from uuid import UUID
 from pydantic import BaseModel
 
-
-def orjson_dumps(v, *, default):
-    return orjson.dumps(v, default=default).decode()
+from .base import BaseOrjsonModel
 
 
-class Genre(BaseModel):
+class Genre(BaseModel, BaseOrjsonModel.Config):
     id: UUID
     name: str
     description: Optional[str] = ''
-
-    class Config:
-        json_loads = orjson.loads
-        json_dumps = orjson_dumps
