+++ service/models/person.py
@@ -1,20 +1,13 @@
 from datetime import date
 
-import orjson
 from uuid import UUID
 from typing import Optional
 from pydantic import BaseModel
 
-
-def orjson_dumps(v, *, default):
-    return orjson.dumps(v, default=default).decode()
+from .base import BaseOrjsonModel
 
 
-class Person(BaseModel):
+class Person(BaseModel, BaseOrjsonModel.Config):
     id: UUID
     full_name: str
     birth_date: Optional[date]
-
-    class Config:
-        json_loads = orjson.loads
-        json_dumps = orjson_dumps
