+++ setup.cfg
@@ -1,4 +1,20 @@
 [flake8]
-exclude = manage.py, settings.py, 01_etl/app/components/*, urls.py,
-per-file-ignores = __init__.py: D104, 01_etl/app/movies/migrations/*: E501, 01_etl/*:W291, W293
+ignore =
+    W503,
+    D401,
+    E501,
+    I001,
+    I005
+exclude =
+    manage.py,
+    settings.py,
+    01_etl/app/components/*,
+    urls.py,
+    tests/,
+    venv/
+per-file-ignores =
+    __init__.py: D104,
+    01_etl/app/movies/migrations/*: E501,
+    01_etl/*:W291,
+    W293
 max-line-length=120
